requirements.txt
Django==4.1.2
djangorestframework==3.14.0
django-taggit==3.0.0
drf-spectacular==0.24.2
easy-thumbnails==2.8.3
model-bakery==1.7.1
psycopg2-binary==2.9.4
pytest==7.1.3
pytest-django==4.5.2
pytest-cov==4.0.0
python-dotenv==0.21.0
django-filter==22.1
Pillow==9.2.0
